+++ modules/geometry/tests/test_rotx.m
@@ -7,15 +7,25 @@
 % SPDX-License-Identifier: LGPL-3.0-or-later
 % LICENCE_BLOCK_END
 %=============================================================================
+arch = computer('arch');
+%=============================================================================
 R = rotx(90);
 REF =    [1     0     0;
 0     0    -1;
 0     1     0];
-assert_isequal(R, REF);
+if strcmp(arch, 'win32')
+  assert_isapprox(R, REF, 1e-15);
+else
+  assert_isequal(R, REF);
+end
 %=============================================================================
 R = rotx(180);
 REF = [     1     0     0;
 0    -1     0;
 0     0    -1];
-assert_isequal(R, REF);
+if strcmp(arch, 'win32')
+  assert_isapprox(R, REF, 1e-15);
+else
+  assert_isequal(R, REF);
+end
 %=============================================================================